refactor to create generic list, so I could do list performance enhancement
