+++ inventory/views/subitems_list_view.py
@@ -1,8 +1,8 @@
-from inventory.views import ItemsListView
+from inventory.views import GenericListView
 from inventory.models import Subitem
 
 
-class SubItemsListView(ItemsListView):
+class SubItemsListView(GenericListView):
     object_type = Subitem
     template = 'inventory/subitem_list.tmpl'
     order_fields = ('item__disposition', 'item__category')
